[TRANSLATION] Course BIZ101 - IT (#1614)
* Update preamble BIZ101_it.md

* Update Chapter 1.1_it.md

Small orthographic and syntax corrections

* Update chapter 2.1_it.md

Small orthographic and syntax corrections, substitution of the word "rete/i" with network when used without complement, formatted in bold and italic some parts of the text

* Update chapter 2.1_it.md

Make letter n with constant spacing to highlight it from the text with `n`.

* Update chapter 2.2_it.md

Fixing tables and inserting text, corrections small errors and formatting bold and italic

* Update chapter 2.3_it.md

Fixes small errors and formatting bold and italic

* Update chapter 2.4_it.md

Fixes small errors and formatting bold and italic

* Update chapter 3.1_it.md

Fixing small errors and replacing bold and italic formatting

* Update preambol_it.md

Change title

* Update chapter 3.2_it.md

Fix some small errors, bold and italic formatting, and table layout.

* Update chapter 3.3_it.md

Fixing some small errors, formatting bold and italic.

* Update chapter 4.1_it.md

Fixing some small errors, formatting bold and italic, and a lot of table arrangement

* Update chapter 4.2_it.md

Fixing some small errors, formatting bold and italic.

* Update chapter 4.3_it.md

Fixing some small errors, formatting bold and italic and substitutions of "b" and "B" in bitcoin and Bitcoin according to context... if seen as coin or protocol

* Update chapter 4.4_it.md

Little change

* Update chapte 4.5_it.md

Little change

* Update chapter 4.6_it.md

Small corrections, bold and italic formatting.

* Update chapter5.1_it.md

Small corrections, bold and italic formatting and substitution letter "b" with "B" and vice versa to diversify the coin from the protocol

* Update chapter 5.2_it.md

Small corrections, bold and italic formatting and substitution letter "b" with "B" and vice versa to diversify the coin from the protocol

* Update chapter 5.3_it.md

Small corrections, bold and italic formatting and substitution letter "b" with "B" and vice versa to diversify the coin from the protocol, use word "satoshi" in lowercase because considered as currency in the chapter

* Create 04.webp

* Delete courses/biz101/assets/notext/04.webp

* Create 0004.webp

* Add 04.webp notext

Add only image

* Delete courses/biz101/assets/notext/0004.webp

* Add 03.webp no text

Add only image

* Add 05.webp no text

Add only image

* Add 06.webp no text

Add only image whit original name

* Delete courses/biz101/assets/notext directory

* Update Chapter 3.1_it.md

Add space

* Update it.md

Final revision of chapters 1 - 2 - 3, small formatting corrections, spelling and tried to distinguish the bitcoin asset from the Bitcoin protocol as accurately as possible.

* Update chapter 4 - 5 - 6_it.md

Final revision of chapters 4 - 5 - 6, small formatting corrections, spelling and tried to distinguish the bitcoin asset from the Bitcoin protocol as accurately as possible.

* Update all chapter_it.md

Review the whole course from the preview, making some small corrections.

* Update 000_it.yml

Remove a space.

* Update 005_it.yml

Replaced "minatori" with miner to maintain a specific English term (in Italian foreign words should not be made plural)

* Update 006_it.yml

Replaced Bitcoin with bitcoin because it is coin.

* Update 007_it.yml

Replaced finality with block confirmation

* Update 009_ it.yml

Entered a term in italic

* Update 015_it.yml

Replaced consumed with employees as the noun had already been used shortly before.

* Update 017_it.yml

Replaced consumed with employees as the noun had already been used shortly before.

* Update 018_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 019_it.yml

Replaced Bitcoin with bitcoin as it indicates value as currency and not as protocol, and removed the definite article when talking about protocol.

* Update 021_it.yml

Replaced Bitcoin with bitcoin as it indicates value as currency and not as protocol.

* Update 025_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 024_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol. (add DCA)

* Update it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 022_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 026_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 027_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 029_it.yml

Replaced lampo with Lightning as it indicates value as coin and not as protocol.

* Update 032_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 033_it.yml

Adjusted the article of Lightning Network

* Update 035_it.yml

Adjusted the article of Lightning Network

* Update it.yml

Adjusted the article of Lightning Network

* Update 036_it.yml

Adjusted the article of Lightning Network

* Update 023_it.yml

* Update 036_it.yml

* Update 036_it.yml

* Update 036_it.yml

Adjusted the article of Lightning Network

* Update 038_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol from coin

* Update 039_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol from coin

* Update 040_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol from coin

* Update 041_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol from coin

* Update 042_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol from coin

* Update 043_it.yml

Deleted item in front of Bitcoin

* Update 054_it.yml

Substitute "passero" with sparrow

* Update 055_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol with coin

* Update 056_it.yml

Replaced "custodiale" with custodial

* Update 057_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol with coin

* Update 059_it.yml

Replaced "custodiale" with custodial and add brain wallet LN

* Update 060_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol with coin

* Update 062_it.yml

Substitution "Nodo aperto" with Open Node

* Update 063_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol with coin

* Update 065_it.yml

Substitution "Nodo apero" with Open Node

* Update 066_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol with coin and sostitution "Nodo aperto" with Open Node

* Update069_ it.yml

Replaced Bitcoin with bitcoin to differentiate protocol with coin and substitution "Nodo aperto" with Open Node

* Update 070_it.yml

Substitution "Cruscotti" con Dashboard

* Update 075_it.yml

Replaced Bitcoin with bitcoin to differentiate protocol with coin

* Update it.md

* Update it.md

Fixed formatting as from original file in English.

* Update BIZ101_it.md

Completed the course review

* Update BIZ_101_quiz009it.yml

Deleted article In front of Lightning Network and fix formatting

* Update it.yml

* Update quizz_022_it.yml

* Update quizz_023_it.yml

Replaced bitcoin which asset in minuscule by diversifying it from the protocol

* Update quizz_026_ it.yml

Replaced bitcoin which asset in minuscule by diversifying it from the protocol

* Update quizz_027_it.yml

Replaced bitcoin which asset in minuscule by diversifying it from the protocol

* Update quizz_038_ it.yml

Replaced bitcoin which asset in minuscule by diversifying it from the protocol

* Update quiz_040_it.yml

minus "u"

* Update quizz_054_it.yml

* Update quizz_055_it.yml

* Fix new format

---------

Co-authored-by: MarJJ <mariannacon@protonmail.com>
Co-authored-by: asi0 <90136504+Asi0Flammeus@users.noreply.github.com>