[PROOFREADING] btc204 + quiz - de (#3246)
* Update de.md

* Update de.md

* Section 2.1

* Update de.md - Section 2.2

* Update de.md - Section 2.3

* Update de.md - Section 3.2

* Update de.md - Section 3.2

* Update de.md - Section 3.3

* Update de.md - Section 3.5

Wallet imprint is hard to translate. AI proposes the following:
„Abdruck im Portemonnaie“ (physisch: Stempel, Abdruck im Leder oder Formabdruck durch Karten)

„Spuren in der Geldbörse“ (im übertragenen Sinn: welche Transaktionen/Finanzmuster sich dort zeigen)

„Wallet-Fingerabdrücke“ (im digitalen Bereich: charakteristische Spuren oder Identifikatoren von Krypto-Wallets)

I decided to use the latter because it seems to nail the meaning best.

* Update de.md - Section 3.5

* Update de.md

* Update de.md

* Update de.md - Section 4.1

* Update de.md - Section 4.2

* Update de.md - Section 4.3

* Update de.md - ections 4.5 and 4.6

txstats.com is not available right now (for me)

* Update de.md - Section 5.1

The titel "Mixer, Coinjoin, Mixer..." contains two times the word "mixer" - i am not sure if this is on purpose. THe english version also has it so I did not adapt it. Please check again

* Update de.md - Section 5.2

* Update de.md  - Section 5.3

* Update de.md - Section 5.4

* Update de.md - Section 5.5 + 5.6

* Update de.md - Section 6.1

* Update de.md - Section 6.2 + 6.3

I would use the original word "Ricochet" instead of the german translation. What do you think?

* Update de.md - Section 6.4

* Update de.md - Section 6.5 + 6.6

* Update de.md - Section 6.7

I am not sure about the formulas (if they are rendered correctly).

* Proofread quiz

* Proofread quiz

* Proofread quiz

* Update de.md - Fix typo (after review)

* Fix typos

* fix tables and minor stuff

---------

Co-authored-by: MarJJ <marjjhodl@proton.me>