[PROOFREADING] CYP201 + QUIZZES - German (#3654)
* Update de.md - 2.2

After "**Funktion $\Sigma_1$:**", there is a problem with formatting - because of the wrongly placed dollar sign, I think. Also, the images and the rest of the text is missing afterwards - I guess, this is fixed by inserting the new line after the dollar sign but please double check. Nevertheless, I proofread the rest of the section from the sources. ;)

* Update de.md - 2.3

* Update de.yml

* Update de.md - 3.x

* Update de.md - 5.1

* Update de.md - 5.2

* Fix typo from 'Ketten-Code' to 'Chain-Code'

* 5.3

* 5.4

* Update de.md

* set all "reviewed" flags to their actual status

* Review rest of de.md

* format fix

* fix formula

---------

Co-authored-by: MarJJ <marjjhodl@proton.me>