[Proofreading]-contribution-section-fr (#3692)
* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

I took the liberty of adding spaces between the text and images for better readability.

* Update fr.md

* Update fr.md

I took the liberty of adding spaces between the text and images for better readability.

* Update fr.md

* Update en.md

I added the ``` that changed the appearance of the English version.

* Update en.md

I added the ``` that changed the appearance of the English version.

* Update fr.md

I took the liberty of adding spaces between the text and images for better readability.

* Update fr.md

* Update fr.md

I took the liberty of adding spaces between the text and images for better readability.

* Update fr.md

* Update fr.md

I took the liberty of adding spaces between the text and images for better readability.

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

Lines 30 and 32 should be checked.

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

* Update fr.md

---------

Co-authored-by: MarJJ <marjjhodl@proton.me>